+++ kv_relation-store.go
@@ -31,6 +31,10 @@ type RelationStore[ID ~[]byte | ~string, RelatedID ~[]byte | ~string] interface 
 	StreamIDs(ctx context.Context, ch chan<- ID) error
 	// StreamRelatedIDs return all existing relationIDs
 	StreamRelatedIDs(ctx context.Context, ch chan<- RelatedID) error
+	// MapIDRelations maps all entry to the given func
+	MapIDRelations(ctx context.Context, fn func(ctx context.Context, key ID, relatedIDs []RelatedID) error) error
+	// MapRelationIDs maps all entry to the given func
+	MapRelationIDs(ctx context.Context, fn func(ctx context.Context, key RelatedID, ids []ID) error) error
 	// Invert returns the same store with flipped ID <-> RelationID
 	Invert() RelationStore[RelatedID, ID]
 }
@@ -55,6 +59,27 @@ func NewRelationStoreFromRelationStoreTx[ID ~[]byte | ~string, RelatedID ~[]byte
 type relationStore[ID ~[]byte | ~string, RelatedID ~[]byte | ~string] struct {
 	relationStoreTx RelationStoreTx[ID, RelatedID]
 	db              DB
+}
+
+
+func (r *relationStore[ID, RelatedID]) MapIDRelations(ctx context.Context, fn func(ctx context.Context, key ID, relatedIDs []RelatedID) error) error {
+	err := r.db.View(ctx, func(ctx context.Context, tx Tx) error {
+		return r.relationStoreTx.MapIDRelations(ctx, tx, fn)
+	})
+	if err != nil {
+		return errors.Wrapf(ctx, err, "view failed")
+	}
+	return nil
+}
+
+func (r *relationStore[ID, RelatedID]) MapRelationIDs(ctx context.Context, fn func(ctx context.Context, key RelatedID, ids []ID) error) error {
+	err := r.db.View(ctx, func(ctx context.Context, tx Tx) error {
+		return r.relationStoreTx.MapRelationIDs(ctx, tx, fn)
+	})
+	if err != nil {
+		return errors.Wrapf(ctx, err, "view failed")
+	}
+	return nil
 }
 
 func (r *relationStore[ID, RelatedID]) Invert() RelationStore[RelatedID, ID] {